runtimes/eoapi/stac/pyproject.toml
[project]
name = "eoapi.stac"
description = "STAC Metadata service for eoAPI."
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Vincent Sarago", email = "vincent@developmentseed.com"},
]
license = {text = "MIT"}
classifiers = [
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: GIS",
]
dynamic = ["version"]
dependencies = [
    "stac-fastapi.pgstac==3.0.0a1",
    "jinja2>=2.11.2,<4.0.0",
    "starlette-cramjam>=0.3,<0.4",
    "importlib_resources>=1.1.0;python_version<'3.9'",
    "psycopg_pool",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "httpx",
]

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool.pdm.version]
source = "file"
path = "eoapi/stac/__init__.py"

[tool.pdm.build]
includes = ["eoapi/stac"]
excludes = ["tests/", "**/.mypy_cache", "**/.DS_Store"]
